Move formDetails to own folder including tests
