+++ designer/client/components/FormDetails/index.ts
@@ -1,0 +1,1 @@
+export { FormDetails } from "./FormDetails";
